cross_compat_engines refactoring and following rubocop fixes
