+++ lib/thread_pool.rb
