+++ test/findme_parse_test.rb
@@ -16,7 +16,7 @@ class ParseTest < Minitest::Test
     full_name_query = 'jane+smith'
     # search_engines = %w[bing yandex yahoo startpage duckduckgo
     #                     baidu searchencrypt gigablast]
-    search_engines  = ["startpage", "bing", "google"]
+    search_engines  = %w[startpage bing google]
 
     search_engines.each do |search_engine|
       query               = SearchClient.new(full_name_query)
@@ -25,18 +25,18 @@ class ParseTest < Minitest::Test
       pr                  = ParseResults.new(query_result.body)
       all_links           = []
       case search_engine
-      when "google"
-        all_links         = pr.parse_google
-      when "bing"
-        all_links         = pr.parse_bing
-      when "startpage"
-        all_links         = pr.parse_startpage
+      when 'google'
+        all_links = pr.parse_google
+      when 'bing'
+        all_links = pr.parse_bing
+      when 'startpage'
+        all_links = pr.parse_startpage
       # SKIPPING YAHOO. XPATH/CSS nokogiri can of worms
       # when "yahoo"
 
       #   all_links         = pr.parse_yahoo
       else
-        raise RuntimeError.new("search engine undefined")
+        raise 'search engine undefined'
       end
 
       # FOR NOW WE LOOK AT RESPONSE FROM THE METHOD PROGRESS
@@ -52,5 +52,4 @@ class ParseTest < Minitest::Test
       assert all_links['direct'].map { |link| link.match(%r{^https://}) }
     end
   end
-
 end
